Renamed folders and added comments
