--- Sources/Lux/Injector/Injector.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-public protocol Injector {
-
-    var languageIdentifiers: Set<String> { get set }
-
-    func inject(in text: String) -> String
-    func injectAttributed(in text: String) -> NSMutableAttributedString
-}
