+++ Sources/Lux/InjectorImplementations/Json/JSONInjector.swift
@@ -1,6 +1,5 @@
 import Foundation
 
-/// Inject strings into a text depending on the configuration or the delegate.
 public final class JSONInjector: BaseInjector<JSONCategory> {
 
     override var defaultLanguageIdentifiers: Set<String> { ["json", "Json", "JSON", "language-json", "language-Json", "language-JSON"] }
